Use md5 for mtime rather than filemtime
Using the file mtime of a generated asset means that the resulting file name will be different every time this function is run.  This causes problems if load balancing across multiple servers as the mtime will be slightly different on each server.  This will mean each server generates an asset file with a different name.  Re-using the md5 from the pre-minified asset should be safe.